\name{compMPD}
\alias{compMPD}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Compare MPD Estimates
}
\description{
  Compare MPD estimates from various model runs, 
  usually the base run and its sensitivities.
}
\usage{
compMPD(dir=".", type="senso", strSpp="SGR", prefix="sgr.coast.",
   rda.name=NULL, vyrs=1933, ryrs=1935:2026, pyrs=2026:2036,
   figs=FALSE, tabs=FALSE, png=FALSE, pngres=400, PIN=c(9,9), lang=c("e"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dir}{\code{character} -- name of directory containing 
     binary output from call to \code{'gatherMPD()'}}
  \item{type}{\code{character} -- type of MPD runs (usually \code{'senso'})}
  \item{strSpp}{\code{character} -- three-letter code of species}
  \item{prefix}{\code{character} -- model prefix for function output
    (e.g., \code{'sgr.coast.'})}
  \item{rda.name}{\code{character} -- input name of binary output from
     call to \code{'gatherMPD()'}}
  \item{vyrs}{\code{numeric} -- virgin biomass year(s) generated by SS3}
  \item{ryrs}{\code{numeric} -- reconstruction years for the population model}
  \item{pyrs}{\code{numeric} -- projection years for the population model}
  \item{figs}{\code{logical} -- if \code{TRUE}, create trajectory figures}
  \item{tabs}{\code{logical} -- if \code{TRUE}, create tables of MPD estimates}
  \item{png}{\code{logical} -- if \code{TRUE}, send trajectory figures to a \code{PNG} file}
  \item{pngres}{\code{numeric} -- resolution (pixels/inch) for \code{.png} output}
  \item{PIN}{\code{numeric} -- plot device size (inches) for width and height}
  \item{lang}{\code{character} -- figure language (\code{'e'} for english, \code{'f'} for french)}
}
\details{
  Compare a set of MPD runs, visually and/or tabularly.
  Created primarily for comparing MPD sensitivities
  to the base run for SGR 2025.
}
\value{
	Figure and/or table files for use in LaTeX 
	(primarily referenced by Sweave \code{.Rnw} code).
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: remote office, Vancouver BC\cr
  Last modified \code{Rd: 2026-04-28}
}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
In package \pkg{PBSsynth}:\cr
  \code{\link[PBSsynth]{gatherMPD}}\cr
In package \pkg{PBStools}:\cr
  \code{\link[PBStools]{changeLangOpts}},
  \code{\link[PBStools]{clearFiles}},
  \code{\link[PBStools]{createFdir}},
  \code{\link[PBStools]{formatCatch}},
  \code{\link[PBStools]{linguaFranca}}\cr
In package \pkg{PBSmodelling}:\cr
  \code{\link[PBSmodelling]{addLegend}},
  \code{\link[PBSmodelling]{expandGraph}},
  \code{\link[PBSmodelling]{lucent}},
  \code{\link[PBSmodelling]{unpackList}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{hplot}
\keyword{file}
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
\concept{appFuns}
% \concept{ ~cpt2 }
% Use only one concept per line.
