\name{plotBh}
\alias{plotBh}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Plot Biomass vs Steepness
}
\description{
  Plot absolute reference biomass (\eqn{B_0}{B0} and \eqn{B_\text{MSY}}{Bmsy}) or
  relative current biomass (\eqn{B_{2026}/B_0}{B2026/B0} and \eqn{B_{2026}/B_\text{MSY}}{B2026/Bmsy})
  against estimated steepness \eqn{h}{h} for samples from MCMC posteriors.
}
\usage{
plotBh(dmcmc, Bmcmc, Bmsy, xlim, xfld="BH_h", 
   ytype="rel", currYr, base, top, hbin=0.015, 
   png=FALSE, pngres=400, PIN=c(10,7.5))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dmcmc}{\code{numeric} -- data frame imported from SS3 output by function \code{r4ss::SSgetMCMC}}
  \item{Bmcmc}{\code{numeric} -- data frame of spawning biomass by year (columns) for each MCMC sample (rows)}
  \item{Bmsy}{\code{numeric} -- vector of MCMC samples of \eqn{B_\text{MSY}}{Bmsy}}
  \item{xlim}{\code{numeric} -- limits of the x-axis}
  \item{xfld}{\code{character} -- field name containing x-values (in this case, steepness estimates)}
  \item{ytype}{\code{character} -- type of data to plot: `abs'=absolute (reference biomass) or
    `rel'=relative (current biomass relative to reference biomass)}
  \item{currYr}{\code{numeric} -- current year}
  \item{base}{\code{numeric} -- lowest value of \code{y} for the y-axis 
    (set to a negative value to allow room for a steepness histogram along the bottom axis)}
  \item{top}{\code{numeric} -- highest value of \code{y} for the y-axis 
    (default's to 0.99 quantile of \code{y} to remove extreme estimates)}
  \item{hbin}{\code{numeric} -- bin width for steepness histogram}
  \item{png}{\code{logical} -- if \code{TRUE}, send the graphical output to a \code{PNG} file}
  \item{pngres}{\code{numeric} -- resolution for the \code{PNG} file in pixels/inch}
  \item{PIN}{\code{numeric} -- dimensions (width, height) of the \code{PNG} file in inches}
}
\details{
  The figure illustrates how dependent estimates of reference points are on the estimate of steepness.
  Often, \eqn{B_\text{year}/B_0}{Byear/B0} remains constant across values of steepness, whereas
  \eqn{B_\text{year}/B_\text{MSY}}{Byear/Bmsy} increases with increasing steepness.
}
\value{
  Returns \emph{nada}, but creates a diagnostic figure.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Remote office, Vancouver BC\cr
  Last modified \code{Rd: 2026-02-20}
}
\note{
  A smooth line through sample points is generated by the function 
  \code{stats::loess.smooth} using default settings.

  At lower estimated steepness values, MSY-based calculations return a non-finite number.
  The plot indicates the lowest steepness (vertical dasded line) that generated a finite MSY value.
}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
In package \pkg{PBSsynth}:\cr
  \code{\link[PBSsynth]{calcRhat}},
  \code{\link[PBSsynth]{plotSS.dmcmc}}\cr
In package \pkg{PBSmodelling}:\cr
  \code{\link[PBSmodelling]{drawBars}}\cr
In package \pkg{stats}:\cr
  \code{\link[stats]{loess.smooth}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{hplot}
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
\concept{plotFuns}
\concept{diagnostics}
% Use only one concept per line.

