\name{plotTraj}
\alias{plotTraj}
\title{
  Plot Trajectory Medians
}
\description{
  Plot the trajectory medians from various MCMC sensitivity runs in one figure.
}
\usage{
plotTraj(sdat, index, traj="B", bdat=NULL, y0=FALSE, lab.stock, 
   startYear=1935, currYear=2025, sen.lab,
   col=c("black","green4","blue","red","purple","orange"), 
   lty=c(2:6), logR=FALSE, 
   png=FALSE, pngres=400, PIN=c(8,8), lang=c("e","f"), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sdat}{\code{list} -- sensitivity runs generated by running Sweave code to 
    compile a Model Results appendix, which gathers run data (currently in: \code{stock[[istock]]$Sens}).
    The object comprises one or more model runs, identified at the first level in the list's hierarchy.}
  \item{index}{\code{numeric} -- vector of indices pointing to runs that a user wishes to compare.
    For example, the index typically points to a base case and a number of sensitivity runs.}
  \item{traj}{\code{character} -- vector of trajectory names (e.g., \code{B} = spawning biomass,
    \code{Bt/B0} = spawning biomass depletion); options are limited.}
  \item{bdat}{\code{list} -- central run of the base case that is used to intitiate sensitivity runs;
    list contains vectors of medians for \code{BtB0} (depletion), \code{R} (recruitment), and \code{U} (harvest rate).
    Information on the central run is found in \code{stock[[istock]]$Base[[run.rwt]]}.}
  \item{y0}{\code{logical} -- if \code{TRUE}, start the y-axis at 0.}
  \item{lab.stock}{\code{character} -- name of stock to appear in the name of the \code{'png'} output file, if \code{png=TRUE}.}
  \item{startYear}{\code{numeric} -- starting year of the model}
  \item{currYear}{\code{numeric} -- current year of the model}
  \item{sen.lab}{\code{character} -- sensitivity names of various trajectories}
  \item{col}{\code{character} -- names of colours for each trajectory.}
  \item{lty}{\code{numeric|character} -- line types for each trajectory.}
  \item{logR}{\code{logical} -- if \code{TRUE}, then log the trajectories (usually of recruitments)}
  \item{png}{\code{logical} -- if \code{TRUE}, send the figure output to a \code{'png'} device.}
  \item{pngres}{\code{numeric} -- resolution (pixels/inch) for \code{'png'} file.}
  \item{PIN}{\code{numeric} -- vector of width and height (inches) for \code{'png'} file.}
  \item{lang}{\code{character} -- a vector of letters that denote the language for output:
    currently only \code{"e"} (english) and \code{"f"} (french).}
  \item{\dots}{\code{list} -- additional arguments for \code{lines} (\code{lty} only) and \code{legend}.}
}
\details{
  Trajectories are currently limited to objects in the \code{currentMCMC} object 
  (\code{c("B","VB","R","U")}) and spawning biomass depletion (\code{"BoverB0"}).
}
\value{
  Assigns a trajectory \code{array} object to the global environment.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Remote office, Vancouver BC\cr
  Last modified \code{Rd: 2024-04-16}
}
\note{
  This function is geared for use after a number of Awatea|SS3 model runs have been completed.
  The steps for creating one model run generally involve five stages:\cr
  (i) reweighting abundance & composition data,\cr
  (ii) MPD fit,\cr
  (iii) MCMC fits,\cr
  (iv) MSY calculations, and\cr
  (v) projections.\cr
  After a number of model runs have been generated, the results can be
  gathered and compared.
}
\seealso{
In package \pkg{PBSsynth}:
  \code{\link[PBSsynth]{gatherMCMC}},
  \code{\link[PBSsynth]{plotSS.senso}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}% use one of  RShowDoc("KEYWORDS")
\concept{plotFuns.r}


