\name{small.funs}
%- Also NEED an '\alias' for EACH other topic documented here.
\alias{allEqual}
\alias{cquantile.vec}
\alias{divTwo}
\alias{findPos}
\alias{gather.vectors}
\alias{getActDim}
\alias{getAssYrs}
\alias{getMPD}
\alias{getNpan}
\alias{getYrIdx}
\alias{is.numStr}
\alias{med5.95}
\alias{medCI}
\alias{ptab}
\alias{qtab}
\alias{relabelTex}
\alias{stab}
\title{
  Small Helper Functions
}
\description{
  Small helper routines to facilitate transactions in PBSsynth.
}
\usage{
allEqual(x)
cquantile.vec(z, prob)
divTwo(x, dig=decdig)
findPos(target, choices)
gather.vectors(run, rwt, ver, use)
getActDim(A)
getAssYrs(spp.code, area.name)
getMPD(obj)
getNpan()
getYrIdx(yrNames, mod=5)
is.numStr(x)
med5.95(xx.MCMC, dig=0, quants3=tcall(quants3))
medCI(x, dig=decdig, CI=quants3[c(1,3)])
ptab(xx)
qtab(xx.MCMC, dig=0, quants3=tcall(quants3))
relabelTex(texinput, prefix, caption)
stab(xx.MCMC, dig=3, quants3=tcall(quants3), print=TRUE)
}
\arguments{
  \item{x}{\code{numeric} -- vector of values (can sometimes be \code{character} also)}
  \item{z}{\code{numeric} -- cumulative quantile of vector}
  \item{prob}{\code{numeric} -- scalar probability}
  \item{dig}{\code{numeric} -- number of decimal places or significant digits, depending on function}
  \item{target}{\code{numeric} -- target scalar value}
  \item{choices}{\code{numeric} -- vector of values}
  \item{run}{\code{numeric} -- run number}
  \item{rwt}{\code{numeric} -- reweight number}
  \item{ver}{\code{character} -- version (e.g. \code{'1a'})}
  \item{use}{\code{logical} -- three element logical vector for [i,j,k] array}
  \item{A}{\code{numeric} -- axis of uncertainty matrix}
  \item{spp.code}{\code{character} -- species 3-letter code}
  \item{area.name}{\code{character} -- area name of stock}
  \item{obj}{\code{numeric} -- MPD object created by model (likely Awatea) support code}
  \item{yrNames}{\code{character} -- years as characters (e.g., \code{'1935'})}
  \item{mod}{\code{numeric} -- used in modulo operation: \code{yrVals \%\% mod==0}}
  \item{xx}{\code{numeric} -- vector of values}
  \item{xx.MCMC}{\code{numeric} -- vector of values from MCMC posterior}
  \item{CI}{\code{numeric} -- vector of three quantiles, usually (0.05, 0.5, 0.95)}
  \item{quants3}{\code{numeric} -- vector of three quantiles, usually (0.05, 0.5, 0.95)}
  \item{texinput}{\code{character} -- tex code generated by `SweaveMPD.Rnw' or `SweaveMCMC.Rnw'}
  \item{prefix  }{\code{character} -- prefix to add to table/figure labels}
  \item{caption }{\code{character} -- prefix to add to table/figure captions}
  \item{print}{\code{logical} -- if \code{TRUE}, print results to current device}
}
%- maybe also 'usage' for other objects documented here.
\details{
  \tabular{ll}{
    \code{allEqual}       \tab determine if all elements in vector equal the first element\cr
    \code{cquantile.vec}  \tab calculate cumulative quantile as a vector\cr
    \code{divTwo}         \tab divide first element of vector by second element\cr
    \code{findPos}        \tab find nearest position of a number in a vector of numbers\cr
    \code{gather.vectors} \tab gather vectors of base run components\cr
    \code{getActDim}      \tab get number of active dimensions from the axis of uncertainty matrix\cr
    \code{getAssYrs}      \tab get modelled current years for previous stock assessments\cr
    \code{getMPD}         \tab get MPD estimated parameters\cr
    \code{getNpan}        \tab get panel number when inside a multi-panel plot\cr
    \code{getYrIdx}       \tab select years for plotting\cr
    \code{is.numStr}      \tab check if strings can be converted to numerics\cr
    \code{med5.95}        \tab print median and credibility envelope, e.g., (0.05, 0.95), to text\cr
    \code{medCI}          \tab print the median and the credible interval\cr
    \code{ptab}           \tab prior tabulation (LaTeX) for lines in table of priors\cr
    \code{qtab}           \tab quantile tabulation summary using decimal places\cr
    \code{relabelTex}     \tab relabel label and caption of table/figure from MPD/MCMC run\cr
    \code{stab}           \tab quantile tabulation summary using significant digits
  }
}
\author{
  \href{mailto:andrew.edwards@dfo-mpo.gc.ca}{Andrew Edwards}, Quantitative Assessment Methods, PBS, Nanaimo BC

  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: remote office, Vancouver BC\cr
  Last modified \code{Rd: 2026-07-16}
}
\keyword{manip}
\concept{smallFuns}

